refactor(scheduler): split scheduler responsibilities
Move scheduler APIs, manager behavior, discovery, execution, worker coordination, types, and process-global runtime state into focused same-package files without changing behavior or ownership.
